fracaso

+++ include/entry.h
@@ -7,7 +7,7 @@
 
 void clock_handler();
 
-void writeMSR(int reg, int val);
+void writeMSR(int reg, unsigned long val);
 
 void syscall_handler_sysenter();
